refactor the old delete excluded term dialog
